+++ src/main.rs
@@ -18,6 +18,7 @@ use serde_json::json;
 use std::cmp::Ordering;
 use std::collections::{BTreeSet, HashMap};
 use std::io::Write;
+use std::time::Instant;
 
 use std::sync::{Arc, Mutex};
 
@@ -58,6 +59,7 @@ fn main() -> Result<()> {
         progress_bars.set_draw_target(ProgressDrawTarget::hidden());
     }
 
+    let global_start = Instant::now();
     info!("Starting osm-lump-ways v{}", std::env!("CARGO_PKG_VERSION"));
 
     let reader = read_progress::BufReaderWithSize::from_path(&args.input_filename)?;
@@ -238,7 +240,7 @@ fn main() -> Result<()> {
             .to_formatted_string(&Locale::en)
     );
 
-    info!("All data has been loaded. Started processing...");
+    info!("All data has been loaded in {}. Started processing...", format_duration(Instant::now() - global_start));
     info!(
         "Starting the breathfirst search. There are {} groups",
         group_wayid_nodes.len()
@@ -387,10 +389,10 @@ fn main() -> Result<()> {
         } else {
 
             trace!("wg:{} splitting the groups into single paths with Dij algorithm... wg.num_nodeids() = {}", way_group.root_wayid, way_group.num_nodeids());
-            let started = std::time::Instant::now();
+            let started = Instant::now();
             let paths = match dij::into_segments(&way_group, &nodeid_pos, args.min_length_m, args.only_longest_n_splitted_paths, &splitter) {
                 Ok(paths) => {
-                    let duration = (std::time::Instant::now() - started).as_secs_f64();
+                    let duration = (Instant::now() - started).as_secs_f64();
                     log!(
                         if paths.len() > 20 || duration > 2. { Debug } else { Trace },
                         "Have generated {} paths from wg:{} ({} nodes) in {:.1} sec. {:.2} nodes/sec",
@@ -638,7 +640,7 @@ fn main() -> Result<()> {
         Ok(()) as Result<()>
     })?;
 
-    info!("Finished");
+    info!("Finished all in {}", format_duration(Instant::now() - global_start));
     Ok(())
 }
 #[derive(Debug, Clone, Hash, serde::Serialize, PartialEq, Eq)]
@@ -769,4 +771,38 @@ fn write_geojson_feature_directly(
 enum OutputFormat {
     GeoJSON,
     GeoJSONSeq,
+}
+
+pub fn format_duration_human(duration: &std::time::Duration) -> String {
+    let sec_f = duration.as_secs_f32();
+    if sec_f < 60. {
+        let msec = (sec_f * 1000.).round() as u64;
+        if sec_f > 0. && msec == 0 {
+            "<1ms".to_string()
+        } else if msec > 0 && duration.as_secs_f32() < 1. {
+            format!("{}ms", msec)
+        } else {
+            format!("{:>3.1}s", sec_f)
+        }
+    } else {
+        let sec = sec_f.round() as u64;
+        let (min, sec) = (sec / 60, sec % 60);
+        if min < 60 {
+            format!("{}m{:02}s", min, sec)
+        } else {
+            let (hr, min) = (min / 60, min % 60);
+            if hr < 24 {
+                format!("{}h{:02}m{:02}s", hr, min, sec)
+            } else {
+                let (day, hr) = (hr / 24, hr % 24);
+                format!("{}d{:02}h{:02}m{:02}s", day, hr, min, sec)
+            }
+        }
+    }
+}
+
+
+fn format_duration(d: std::time::Duration) -> String
+{
+    format!("{} ( {:>.1}sec )", format_duration_human(&d), d.as_secs_f32())
 }