+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## v0.12.0 (2023-10-19)
 
 * Output improvements with log messages and progress bars