+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * More options to filter the data (`--only-these-way-groups-nodeid`) to track
   down debugging
 * `all_wayids` option now a string `wNNN` which can be fed into `osmium getid`.
+* Fix bug where it crashed if a way has a doubled node (happened in Antarctica)
 
 ## v0.12.0 (2023-10-19)